+++ RELEASENOTES.md
@@ -46,6 +46,7 @@
     *   MP3: Fix `Searched too many bytes` error by correctly ignoring trailing
         non-MP3 data based on the length field in an `Info` frame
         ([#1480](https://github.com/androidx/media/issues/1480)).
+    *   Add option to enable index-based seeking in `AmrExtractor`.
 *   DataSource:
     *   Update `HttpEngineDataSource` to allow use starting at version S
         extension 7 instead of API level 34